feat: block collabs
